requirements.txt
discord.py==2.1.0
gallery_dl==1.24.5
python-dotenv==0.21.1
rich==13.3.1
discord-webhook==1.0.0
gppt==3.1.0